fix: lint and refactor
